requirements.txt
pyyaml
dnspython
fastapi
uvicorn
requests
python-dotenv
python-whois
requests
tldextract
FastAPI
dnspython
rich